221105 识别完整流程
TODO: 添加对中文识别的支持
