+++ .idea/workspace.xml
@@ -77,7 +77,7 @@
       <workItem from="1667182601400" duration="1066000" />
       <workItem from="1667201429221" duration="60320000" />
       <workItem from="1667373295116" duration="42425000" />
-      <workItem from="1667554000263" duration="15440000" />
+      <workItem from="1667554000263" duration="26605000" />
     </task>
     <servers />
   </component>
@@ -97,6 +97,6 @@
   </component>
   <component name="com.intellij.coverage.CoverageDataManagerImpl">
     <SUITE FILE_PATH="coverage/LicensePlateRecognition$main.coverage" NAME="main Coverage Results" MODIFIED="1667182763392" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$/src" />
-    <SUITE FILE_PATH="coverage/LicensePlateRecognition$LicensePlateRecognition.coverage" NAME="LicensePlateRecognition Coverage Results" MODIFIED="1667576985851" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$/src" />
+    <SUITE FILE_PATH="coverage/LicensePlateRecognition$LicensePlateRecognition.coverage" NAME="LicensePlateRecognition Coverage Results" MODIFIED="1667621287486" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$/src" />
   </component>
 </project>
